+++ src/Introspection/App/Ui5Target.php
@@ -18,9 +18,19 @@ namespace LaravelUi5\Core\Introspection\App;
 final readonly class Ui5Target
 {
     public function __construct(
-        public string $key,
-        public string $name,
+        private string $key,
+        private string $name,
     )
     {
+    }
+
+    public function getKey(): string
+    {
+        return $this->key;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
     }
 }